feat: refactoring controller
