+++ tsconfig.json
@@ -17,7 +17,7 @@
   },
   "include": [
     "src/**/*.ts"
-  ],
+, "src/balancer.ts"  ],
   "exclude": [
     "node_modules",
     "dist"
